pyproject.toml
[project]
name = "crm"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-admin-charts>=1.5.0",
    "django-ckeditor-5>=0.2.18",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.1",
    "django-redis>=6.0.0",
    "django-rest-knox>=5.0.2",
    "djangorestframework-stubs>=3.16.6",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
